requirements.txt
appdirs==1.4.4
astor==0.8.1
asttokens==2.4.1
boto3==1.33.13
botocore==1.33.13
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
distro==1.8.0
executing==2.0.1
fonttools==4.46.0
greenlet==3.0.2
idna==3.6
ipython==8.18.1
jedi==0.19.1
jmespath==1.0.1
JPype1==1.4.1
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pandasgui==0.2.14
parso==0.8.3
Pillow==10.1.0
plotly==5.18.0
prompt-toolkit==3.0.42
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyarrow==14.0.1
Pygments==2.17.2
pynput==1.7.6
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
qtstylish==0.1.5
requests==2.31.0
s3transfer==0.8.2
six==1.16.0
SQLAlchemy==2.0.23
stack-data==0.6.3
tabula==1.0.5
tabula-py==2.9.0
tenacity==8.2.3
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.12
wordcloud==1.9.3
